Add "--debug" flag to compression
The program is now able to save polynomial compressed datastreams in two
formats: the old one is now the "debug" format, which adds some overhead
to the space required by the data. The new format is more compact and
relies to the "encode" and "decode" functions implemented in the latest
release of "libpolycomp".