Improve the "smart" optimizer
The smart optimizer now allows to specify the initial radius of the ring used
in the optimization (via the parameters `opt_delta_coeffs` and
`opt_delta_samples`), as well as the maximum number of rings to test (via
`opt_max_num_of_iterations`). The procedure is now smart enough to be able to
avoid local minima by carefully crafting these parameters.

Internally, the mesh grid generation code has been generalized enough to be
usable with both optimization strategies, the "exhaustive" method and the
"smart" method. (Previously it was only able to save meshes for the exhaustive
method.)